+++ components/button/index.js
@@ -1,7 +1,6 @@
 import Button from './Button';
 import ButtonGroup from './ButtonGroup';
 import IconButton from './IconButton';
-import LinkButton from './LinkButton';
 
 export default Button;
-export { Button, ButtonGroup, IconButton, LinkButton };
+export { Button, ButtonGroup, IconButton };