+++ components/index.js
@@ -2,7 +2,7 @@ import Avatar, { AvatarStack } from './avatar';
 import Badge from './badge';
 import Banner from './banner';
 import Box from './box';
-import Button, { ButtonGroup, IconButton, LinkButton } from './button';
+import Button, { ButtonGroup, IconButton } from './button';
 import { Heading1, Heading2, Heading3, Heading4, Monospaced, TextBody, TextDisplay, TextSmall } from './typography';
 import Bullet from './bullet';
 import Checkbox from './checkbox';
@@ -68,7 +68,6 @@ export {
   IslandGroup,
   Label,
   Link,
-  LinkButton,
   LoadingMolecule,
   LoadingSpinner,
   Menu,